# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Printa Ecosystem API
  version: 1.0.0
  description: |
    The Printa Ecosystem API supports vendor onboarding, stores, catalogues, orders,
    routing, production, point of sale, billing, payments, notifications, and external
    communications. All timestamps are RFC 3339 UTC strings and all identifiers are UUIDs
    unless an endpoint documents otherwise.

    Protected endpoints require an `Authorization: Bearer <JWT>` header. Roles are enforced
    by the backend; an operation description identifies any additional role restriction.
  contact:
    name: Printa Engineering
servers:
  - url: /
    description: Current environment
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceOrder' }
      responses:
        '201': { $ref: '#/components/responses/ObjectCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/orders/{id}:
    parameters: [ { $ref: '#/components/parameters/ID' } ]
    get:
      tags: [Orders]
      summary: Get an order by identifier
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Orders]
      summary: Cancel an order
      responses:
        '204': { description: Order cancelled }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/orders/number/{number}:
    parameters: [ { $ref: '#/components/parameters/OrderNumber' } ]
    get:
      tags: [Orders]
      summary: Get an order by human-readable order number
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/orders/{id}/status:
    parameters: [ { $ref: '#/components/parameters/ID' } ]
    patch:
      tags: [Orders]
      summary: Transition an order status
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusUpdate' }
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/orders/store/{store_id}:
    parameters: [ { $ref: '#/components/parameters/StoreID' } ]
    get:
      tags: [Orders]
      summary: List orders for a store
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/ArrayResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/orders/customer/{customer_id}:
    parameters: [ { $ref: '#/components/parameters/CustomerID' } ]
    get:
      tags: [Orders]
      summary: List orders for a customer
      responses:
        '200': { $ref: '#/components/responses/ArrayResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/routing/decisions/order/{order_id}:
    parameters: [ { $ref: '#/components/parameters/OrderID' } ]
    get:
      tags: [Routing]
      summary: Get a routing decision for an order
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/production/jobs/order/{order_id}:
    parameters: [ { $ref: '#/components/parameters/OrderID' } ]
    get:
      tags: [Production]
      summary: Get the production job for an order
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/pos/transactions/order/{order_id}:
    parameters: [ { $ref: '#/components/parameters/OrderID' } ]
    get:
      tags: [POS]
      summary: Get POS transactions for an order
      responses:
        '200': { $ref: '#/components/responses/ArrayResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/admin/orders:
    get:
      tags: [Administration]
      summary: List orders across the platform
      x-required-roles: [ADMIN]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/ArrayResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/orders/{id}:
    parameters: [ { $ref: '#/components/parameters/ID' } ]
    get:
      tags: [Administration]
      summary: Get an order across the platform
      x-required-roles: [ADMIN]
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    StoreID:
      name: store_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CustomerID:
      name: customer_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderID:
      name: order_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderNumber:
      name: number
      in: path
      required: true
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    ObjectResponse:
      description: Successful object response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericObject' }
    ArrayResponse:
      description: Successful collection response
      content:
        application/json:
          schema:
            type: array
            items: { $ref: '#/components/schemas/GenericObject' }
    ObjectCreated:
      description: Resource created
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericObject' }
    BadRequest:
      description: Invalid request body, query parameter, or state transition
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid, expired, or inactive JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated caller lacks the required role or resource access
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    GenericObject:
      type: object
      additionalProperties: true
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    CartItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1 }
        unit_price: { type: number, format: double, minimum: 0 }
        configuration:
          type: object
          additionalProperties: true
    PlaceOrder:
      type: object
      required: [store_id, channel, items]
      properties:
        store_id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        channel: { type: string, enum: [ONLINE, POS] }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/CartItem' }
        notes: { type: string }
        delivery_address:
          type: object
          description: Pickup details, a saved customer delivery location, or a validated one-time delivery address.
          properties:
            method: { type: string, enum: [pickup, delivery] }
            store_id: { type: string, format: uuid }
            location_id: { type: string, format: uuid, description: Optional saved customer location. }
            recipient_name: { type: string }
            recipient_phone: { type: string }
            address_line1: { type: string }
            address_line2: { type: string }
            city: { type: string }
            country: { type: string }
            latitude: { type: number, format: double, minimum: -90, maximum: 90 }
            longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        discount: { type: number, format: double, minimum: 0 }
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
...